+++ gulpfile.js
@@ -43,7 +43,6 @@ var mocha = require( "gulp-spawn-mocha" );
 gulp.task( "mocha", function() {
 	return gulp.src( [ "spec/**/*.spec.js" ], { read: false } )
 		.pipe( mocha( {
-			require: [ "spec/helpers/node-setup.js" ],
 			reporter: "spec",
 			colors: true,
 			inlineDiffs: true,
@@ -56,3 +55,5 @@ gulp.task( "watch", function() {
 	gulp.watch( "src/**/*", [ "default" ] );
 	gulp.watch( "{lib,spec}/**/*", [ "mocha" ] );
 } );
+
+gulp.task( "continuous", [ "mocha", "watch" ] );