+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### 0.0.2
  * Add support for custom headers at client and request level
+ * Add support for query parameters
  * Add support for request
  * Update README adapters section
  * Bug #6 - allow halon to handle resource list responses